Use condition variable to wait when sync request queue is full.
Previously, in the (hopefully) rare case that we need to wait for the
checkpointer to create space in the sync request queue, we'd enter a
sleep/retry loop.  Instead, create a condition variable so the
checkpointer can wake us up whenever there is a transition from 'full'
to 'not full'.

Discussion:  https://postgr.es/m/20220226213942.nb7uvb2pamyu26dj%40alap3.anarazel.de