Add logging for ProcSignalBarrier mechanism.
To help with diagnosis of systems that are not processing
ProcSignalBarrier requests promptly, add a LOG message every 5 seconds
if we seem to be wedged.  Although you could already see this state via
wait events in pg_stat_activity, the log message also shows the PID of
the process that is preventing progress.

Also add DEBUG1 logging around the whole loop.