Add AIX pollset support to WaitEventSet API.
This AIX stateful polling interface is similar to the epoll and kqueue
interfaces on other operating systems, and allows the kernel to poll a
set of descriptors efficiently.  As with the other interfaces, this is
expected to improve performance on large systems by reducing contention,
most obviously on the postmaster pipe.

After this change, there is only one operating system left in our
support list where we use plain old poll() for WaitEventSetWait():
Solaris.

XXX Compiles and survives light testing on AIX 7.1.  Could be made
slightly better on AIX 7.3.  I no longer have access to big fast NUMA
AIX machines to test the performance improvement, so never proposed to
the mailing list.

XXX No AIX users seem to be left on the mailing list.  So probably there
is no one who could test, review and polish this patch.  It is
approximately 2 decades too late.  But if someone out there wants it...