Remove configure probe for sys/uio.h.
<sys/uio.h> is in SUSv2, and all targeted Unix system have it, so we
might as well drop the probe (in fact we never really needed this one,
as even HP-UX 10.20 had it).  It's where struct iovec is defined, and as
a common extension, it's also where non-standard preadv() and pwritev()
are declared on systems that have it.

It's a little confusing so let's spell what our pg_iovec.h header does
for the OSes in the build farm today:

Windows: defines our own struct and declares our own functions
Solaris: includes sys/uio.h for the struct and declares our own functions
Every other targeted Unix: includes sys/uio.h for the struct and the functions