Doc: Corrections to claims about wal_sync_method.
We claimed that fsync_writethrough called fsync() and then flushed write
caches, but that is untrue.  Describe the actual system call, and
mention that it's macOS-only.

We claimed that all the wal_sync_method values were equally reliable,
except that fsync_writethrough might flush write caches.  In fact
open_[data]sync might differ from f[data]sync on several systems in
terms of power loss risk, depending on whether the filesystem tries to
use FUA, whether the driver passes it on, and whether the device honors
it.  That's too technical a topic to go into here, but let's remove the
misleading statement.

We talked about "preventing caching" when we meant "flushing caches".

We said that open_[data]sync would get you O_DIRECT where available, but
that is not generally true.  It is true only with wal_level=minimal and
max_wal_senders=0, which must surely be a very rare configuration.