Make ConditionVariableCancelSleep() more efficient.
Commit b91dd9de worried about the case where a single wakeup from
ConditionalVariableSignal() (as opposed to ConditionVariableBroadcast())
could be lost if the selected backend decided to cancel its sleep.
That commit's solution was to forward the signal to another backend when
canceling.

Unfortunately this can lead to a lot of unnecessary extra signals in the
more common case of broadcasts, which doesn't have the lost signal
problem.

Changes the API so that ConditionVariableCancelSleep() also reports
whether you were signaled, so that hypothetical clients that are worried
about that problem can handle it themselves.