Silence new missing contrecord error.
Commit adb46615 added an explicit error message when a contrecord was
found to be missing.  It turns out that the new message can generate
some log noise during logical replication walsender shutdown.  That
could probably be fixed other ways, but for now, partially revert that
change.  It's enough to record that an error is queued (meaning that the
XLogPrefetcherNextRecord() will return NULL + whatever error message is
already queued).

Noticed in the ...