Fix sharedtuplestore.c corruption.
If the final chunk of an oversized tuple being written out to disk was
exactly 32760 bytes, it would be corrupted due to a fencepost bug.

Bug #17619.  Back-patch to 11 where the code arrived.

While testing that, I (tmunro) noticed that the per-participant page
counter was not initialized to zero as it should have been; that wasn't
a live bug when it was written since DSM memory was originally always
zeroed, but since 14 min_dynamic_shared_memory might be configured and
it supplies non-zeroed memory, so that is also fixed here.

Author: Dmitry Astapov <dastapov@gmail.com>
Discussion: https://postgr.es/m/17619-0de62ceda812b8b5%40postgresql.org