WIP Reload collations when versions change.
After ALTER DATABASE|COLLATION * REFRESH [COLLATION] VERSION, start
using the new version immediately in the current transaction.  In the
case of COLLATION objects, this is done by invalidating the cache.  In
the case of the database default, this is done by opening the new
version immediately.

XXX If the transaction aborts, we need to undo the change!
XXX If the transaction commits, should we tell other backends to reload?
XXX This isn't finished yet