Give the postmater a "robust" latch.
A "robust" latch is like a shared latch in that another process can set
it, but it doesn't have any state outside the kernel.  On Unix, we just
send a signal on the sending side and use one of the race-free
mechanisms on the receiving side (self-pipe, signalfd or kqueue so state
exists as unconsumed bytes or pending signals).  On Windows, we use an
Windows event directly.  Robust latches don't have our usual
optimizations to avoid entering the kernel if already set or the other
side isn't waiting, but that means they can't become confused if shared
memory is trashed by a badly behaved process.

On regular Unix builds, the postmaster's latch is copied into child
processes without anything special being required, allowing the child
to set it (that is, send a signal).  On EXEC_BACKEND builds, the copy
has to be initialized via a special new interface using a special cookie
passed between postmaster and child representing the serialized
information required to wake this latch.

The SIGUSR1 handler still exists in the postmaster for receiving promote
and logrotate requests from pg_ctl.  Changes:

1.  Regular backends can send "pmsignals" to eg request parallel workers
with a simple wakeup instead of the two-step SIGUSR1 -> SIGURG wakeup
sequence.

2.  The postmaster no longer needs to stat() "promote" and "logrotate"
every time pmsignals arrive.

3.  We remove a dependency on the signal emulation layer on Windows.

XXX WIP