Retire PG_SETMASK() macro.
In the 90s we needed to deal with computers that still had the
pre-standard signal masking APIs.  That's hasn't been necessary for a
very long time on Unix systems (and since c94ae9d8 on Windows system).
The wrapper doesn't expose all the functionality, it's not part of the
API that extensions are expected to be using (or if they are, the change
will be trivial) and it's generally better to use standardised
interfaces.  So, drop the macro and open-code its expansion.