Invent "robust" latches.
A robust latch can wake another process, but doesn't access shared
memory and thus skips the shared memory-based optimizations.  The
processes involved have their own copy, and the latch acts as a wrapper
for the underlying kernel handle.

On Unix, it's a plain old signal sending and receiving mechanism, and
(conceptually) it generates a set event when *any* latch is set, since
the signal is overloaded and shared by all latches; it also
(conceptually) automatically resets.

On Windows, it's a wrapper for a kernel event handle.