Report latches with higher priority than other events.
If a latch is already set, WaitEventSetWait() returns an event
immediately without considering anything else, so latches already had
higher priority in that case.

If we enter the kernel to read events and more than one event is ready
to be reported already, in general there are no ordering guarantees
(that depends on which kernel primitive we're using, among other
complications), but there is one case that deserves special treatment:
if a latch was set in the narrow window between the above check and
entering the kernel to wait (for example, by a signal handler), then
we'd like to have a way to report the latch first.  This makes a rare
race case behave like the common cases.

The solution in this commit is simply to insert WL_LATCH_SET before any
other events reported in the output array, which relies on the caller
providing enough output space for all possible events.

The situation on Windows is a little different, since its kernel only
reports one event at a time, and in a known order based on array
position.  We could make our Windows behavior match Unix by continuing
to drain events in a loop until none are reported, but that might have
other complications and for now is not done.

Therefore, a caller who wants to see latches with higher priority in all
cases should make sure the output array is big enough for all possible
events (for Unix), and also make sure the latch was added to the
WaitEventSet first (for Windows).

This should fix the problem indicated by malleefowl's one-off failure,
where a test did kill(postmaster, SIGHUP), and then connect(), and
assumed that the reload would be handled before the new connection.

That assumption held before commit 7389aad6 because the SIGHUP handler
did the reload work directly, but now it just sets a latch, and with
diabolical scheduling it became possible for the connection to be
created before the handler runs.

Reported-by: Hou Zhijie <houzj.fnst@fujitsu.com>
Reported-by: Tom Lane <tgl@sss.pgh.pa.us>
Discussion: https://postgr.es/m/OS0PR01MB57163D3BF2AB42ECAA94E5C394C29%40OS0PR01MB5716.jpnprd01.prod.outlook.com