Introduce subprocess module to replace system().
Provide a convenient way to handle latch wakeups while waiting for a
subprocess to finish.

Use this to fix the startup process's SIGQUIT handler, which previously
assumed it was OK to call proc_exit() from inside the handler if we were
currently running a restore command.  It was not.  The new
wait_for_latch_or_subprocess() function can be woken up by the SIGQUIT
handler, so that we can safely run the exit from the regular user
context.

XXX Experimental code!  We probably also want a popen-style variant too
(ie with in/out/err pipes).