Don't allow restart points in crash recovery.
Commit 7ff23c6 started the checkpointer and bgwriter during crash
recovery, but failed to consider that a restart point might be acted on
in this state, and cannot work.  For one thing, the SLRUs are not
yet ready, as discovered by an assertion failure, and there are likely
other more complicated problems.

Typically, only one checkpoint record is logged and the checkpointer
ignores it; even if checkpoint_timeout or max_wal_size are reached, it
doesn't have a redo point greater than the control file's, so it is
skipped by CreateCheckPoint().  (XXX: Right?)

However, in the case of recovery from a basebackup taken from a large
and busy system that logged multiple checkpoints during the backup, the
likelihood increases that the checkpointer chooses to act on the restart
point, and then it fails.

Simply decline to record a restart point while in RECOVERY_STATE_CRASH.

XXX WIP

Back-patch to 15, where commit 7ff23c6 arrived.

Reported-by: Bowen Shi <zxwsbg12138@gmail.com>
Analyzed-by: Kyotaro Horiguchi <horikyota.ntt@gmail.com>
Analyzed-by: Andres Freund <andres@anarazel.de>
Analyzed-by: Michael Paquier <michael@paquier.xyz>
Discussion: https://postgr.es/m/17744-2c95e2b7783d7232%40postgresql.org