Use nanosleep() to implement pg_usleep().
The previous coding based on select() required a lot of commentary about
historical portability concerns.  We can just get rid of all that and
use the POSIX nanosleep() function.  The newer clock_nanosleep() variant
might be better, because it can take a CLOCK_MONOTONIC argument to avoid
the chance that ntpd adjustments affect sleep times, but macOS hasn't
caught up with that yet.