Refactor Windows socket FD_CLOSE handling.
Previously, we created Windows event handles for each socket in each
WaitEventSet.  Since FD_CLOSE is reported only once when the remote end
shuts down gracefully, we could lose track of it in rare scenarios where
PostgreSQL relies on WL_SOCKET_READABLE being level-triggered, or not
report it at all when a socket was used with multiple WaitEventSets
consecutively, eg the temporary ones created by WaitLatchOrSocket().

We got away with this in the past when the thing on the other end of the
socket was another PostgreSQL server (ie via postgres_fdw, replication
etc), because the remote server would exit without shutting down or
closing its socket, and that produces a repeating 'abortive' FD_CLOSE.

When commit 6051857fc tried to fix the related problem of random lost
error messages from terminating backends (a source of random BF failures
and user complaints), we discovered this problem.  We had graceful
shutdowns and nice error messages, but we started seeing hangs because
FD_CLOSE could be lost.  Commit 29992a6a5 reverted that change, pending
further investigation of socket event tracking.

This commit addresses the underlying problem as follows:

* for each socket, we now create just one event handle to be used by
  all WaitEventSet objects that are interested in the socket

* for each socket, we now remember whether FD_CLOSE has been reported
  already, so that we can make WL_SOCKET_READABLE sticky, even across
  multiple WaitEventSet objects

The lifetime management of the new per-socket resources is a little
tricky.  For sockets created and closed by regular backend code we could
just manage resource inside our wrapper functions, but we also want
WaitEventSet to work with any sockets from other sources, and especially
sockets created by libpq.  So we allow any socket to be 'blessed', and
we periodically garbage-collect sockets that have been closed such that
there can't be more than 10 useless event handles per process.