Skip the 004_io_direct.pl test if a pre-flight check fails.
Previously the test had a list of OSes that direct I/O was expected to
work on, with the idea that we could easily adjust that if problems
showed up on less common systems.  That worked well enough for all the
systems in our build farm, but didn't survive contact with the Debian
build bots running on tmpfs via overlayfs.

The new approach is to try to create a file with O_DIRECT from perl.  If
that fails, we'll skip the test but report what the error was.

Reported-by: Christoph Berg <myon@debian.org>
Reviewed-by: Dagfinn Ilmari Mannsåker <ilmari@ilmari.org>
Reviewed-by: Andrew Dunstan <andrew@dunslane.net>
Discussion: https://postgr.es/m/ZDYd4A78cT2ULxZZ%40msg.df7cb.de