Assert that pgoff_t is wide enough.
On Windows, we know it's wide enough because we define it directly ourselves.
On Unix, we use off_t, which may only be 32 bits wide on some systems,
depending on compiler switches or macros.  Make absolutely certain that we are
not confused on this point with an assertion, or we'd corrupt large files.