Use ReadRecentBuffer() for btree root page.
The root page of a btree is accessed on every index scan, so it gets
very hot.  We can measure a speed-up on many workloads by pinning it
with ReadRecentBuffer() instead of ReadBuffer(), after remembering where
it was last time in the AM-private cache space in rel->rd_amcache.

Rearrange the existing use of rd_amcache into a new struct
BTAMCacheData.  It's likely that we'll find more things to put in there
in future work.

Discussion: https://postgr.es/m/20230627020546.t6z4tntmj7wmjrfh%40awork3.anarazel.de