Provide PrefetchBuffers() for merged prefetching.
To allow for multiple buffers to be prefetched with operating system
hints, provide a version of PrefetchBuffer() that takes an nblocks
argument.  This can be used to initiate prefetching of a block range
that will later be read.  Results are now returned in an output array.
The traditional PrefetchBuffer() call is now implemented in terms of the
new function.