Provide PrepareReadBuffer() and CompleteReadBuffers().
PrepareReadBuffer() and CompleteReadBuffers() are roughly equivalent to
ReadBuffer(), except that the physical read for multiple adjacent blocks
can be merged into one system call.

CompleteReadBuffers() builds on commit 12f3867f, which allowed a backend
to have multiple buffers marked BM_IO_IN_PROGRESS at the same time.  It
tries to build ranges of consecutive-on-disk blocks to read at once with
smgrreadv().  In between the calls to PrepareReadBuffers() and
CompleteReadBuffers(), any backend is free to perform the read for any
block in that range, in which case CompleteReadBuffers() will see that
it doesn't have to do anything.