Provide StartReadBuffers() and CompleteReadBuffers().
StartReadBuffers() and CompleteReadBuffers() are roughly equivalent to a loop
over PrefetchBuffer() and then ReadBuffer(), for a contiguous range of
blocks.  They allow smgrprefetch() and smgrpreadv() calls to be
consolidated for neighboring blocks.

StartReadBuffers() pins all the buffers it can in the requested range
(which may be fewer than requested due to LimitAdditionalPins()), but
unlike a regular ReadBuffer() it only performs smgrprefetch() for
misses, leaving the read to be performed by CompleteReadBuffers().  (In
later work, it could start true asynchronous I/O).

CompleteReadBuffers() builds on commit 12f3867f, which allowed a backend
to have multiple buffers marked BM_IO_IN_PROGRESS at the same time.  It
tries to build ranges of consecutive-on-disk blocks to read at once with
smgrreadv().  In between the calls to StartReadBuffers() and
CompleteReadBuffers(), any backend is free to perform the read for any
block in that range, in which case CompleteReadBuffers() will see that
it doesn't have to do anything.